+++ wger/nutrition/api/endpoints.py
@@ -1,3 +1,4 @@
 IMAGE_ENDPOINT = 'ingredient-image'
 INGREDIENTS_ENDPOINT = 'ingredient'
+INGREDIENTS_SYNC_ENDPOINT = 'ingredient-sync'
 INGREDIENT_BULK_EXPORT_PATH = 'ingredients/ingredients.jsonl.gz'